+++ src/common/utils/index.js
@@ -1,6 +1,5 @@
 export * from './calculateTotalAmount';
 export * from './getAccountingCodeOptions';
-export * from './getApproveErrorMessage';
 export * from './getOptions';
 export * from './getUserName';
 export * from './invoiceUtils';