+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Error message: Budget Expense Class not found. Refs UINV-198.
+
 ## [2.2.1]https://github.com/folio-org/ui-invoice/tree/v2.2.1) (2020-10-09)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v2.1.3...v2.2.1)